abstract constants and the server
